Fix kinematic-object check float32 quantization
The fixed_objects_remained_kinematic check compared float32 quaternions
with quaternion_geodesic_distance_wxyz.  An otherwise identical float32
quaternion dot product can round below one, so 2*acos reported a constant
6.9e-4 rad even when the kinematic body never moved, failing the 1e-5 rad
gate for every scene with a supported-by (kinematic) object.

Promote the placed/settled states to float64 before measuring, matching
the relative-pose invariant check that already documents the same
quantization.  Measured rotation drift drops from 6.9e-4 rad to ~3e-7 rad
and the check now passes for the person-on-chair cam_22 scene.

Verification:
- python -m py_compile scripts/isaaclab_real_global_cem.py
- cam_22 full pipeline: global CEM passed (reward -1.60), replay passed
  with all 6 objects stable (max displacement 0.0152 m).

+++ scripts/isaaclab_real_global_cem.py
@@ -759,11 +759,18 @@ def _run():
                 for name, values in per_object.items()
             }
         if len(FIXED_NAMES):
+            # Promote to float64: in float32 an otherwise identical quaternion
+            # dot product can round below one, so 2*acos reports ~6.9e-4 rad
+            # even when the kinematic body never moved (see the relative-pose
+            # check below for the same quantization).
             fixed_position_error = torch.linalg.vector_norm(
-                settled[:, fixed_indices, :3] - placed[:, fixed_indices, :3], dim=-1
+                settled[:, fixed_indices, :3].to(dtype=torch.float64)
+                - placed[:, fixed_indices, :3].to(dtype=torch.float64),
+                dim=-1,
             ).max()
             fixed_rotation_error = quaternion_geodesic_distance_wxyz(
-                settled[:, fixed_indices, 3:7], placed[:, fixed_indices, 3:7]
+                settled[:, fixed_indices, 3:7].to(dtype=torch.float64),
+                placed[:, fixed_indices, 3:7].to(dtype=torch.float64),
             ).max()
             maximum_fixed_position_error = max(
                 maximum_fixed_position_error, float(fixed_position_error.item())
